int16

+++ DAOT_mnist/main.py
@@ -67,7 +67,7 @@ def loss_fn(model, features, config, training):
     model_output = model(inputs, training=training)
     
     classification_loss = tf.losses.binary_crossentropy(
-        tf.one_hot(label, axis=-1, depth=config.num_classes, dtype=tf.int32),
+        tf.one_hot(label, axis=-1, depth=config.num_classes, dtype=tf.int16),
         model_output, from_logits=False)
     mean_classification_loss = tf.reduce_mean(classification_loss)
